test: Ensure taskrunner tests can compile on non-linux OS.
There are enterprise tests do not include an OS build restriction
but pull in code that is limited to Linux. This meant the test
could not be run on non-linux and caused IDE errors. Along with
fixing that error, the change consolidates two task runner test
files which were both limited to Linux but used different build
tagging mechanisms.
